rename namespace of servicepatterns to _servicepatterns
