+++ sdk/ai/azure-ai-agents/apiview-properties.json
@@ -6,8 +6,6 @@
         "azure.ai.agents.models.AgentsResponseFormat": "Azure.AI.Agents.AgentsResponseFormat",
         "azure.ai.agents.models.AgentThread": "Azure.AI.Agents.AgentThread",
         "azure.ai.agents.models.AgentThreadCreationOptions": "Azure.AI.Agents.AgentThreadCreationOptions",
-        "azure.ai.agents.servicepatterns.models.AgentV1Error": "Azure.AI.Agents.ServicePatterns.AgentV1Error",
-        "azure.ai.agents.servicepatterns.models.AgentV1ErrorError": "Azure.AI.Agents.ServicePatterns.AgentV1Error.error.anonymous",
         "azure.ai.agents.models.AISearchIndexResource": "Azure.AI.Agents.AISearchIndexResource",
         "azure.ai.agents.models.ToolDefinition": "Azure.AI.Agents.ToolDefinition",
         "azure.ai.agents.models.AzureAISearchToolDefinition": "Azure.AI.Agents.AzureAISearchToolDefinition",
