+++ sdk/ai/azure-ai-agents/azure/ai/agents/_servicepatterns/aio/operations/_patch.py
