+++ sdk/ai/azure-ai-agents/azure/ai/agents/_servicepatterns/models/__init__.py
