+++ sdk/ai/azure-ai-agents/azure/ai/agents/_servicepatterns/models/_patch.py
