+++ sdk/ai/azure-ai-agents/azure/ai/agents/_servicepatterns/operations/__init__.py
