+++ sdk/ai/azure-ai-agents/azure/ai/agents/_servicepatterns/operations/_operations.py
