+++ sdk/ai/azure-ai-agents/azure/ai/agents/aio/_client.py
@@ -15,7 +15,7 @@ from azure.core.pipeline import policies
 from azure.core.rest import AsyncHttpResponse, HttpRequest
 
 from .._utils.serialization import Deserializer, Serializer
-from ..servicepatterns.aio.operations import ServicePatternsOperations
+from .._servicepatterns.aio.operations import ServicePatternsOperations
 from ._configuration import AgentsClientConfiguration
 from .operations import (
     AgentsClientOperationsMixin,
